added more command tests, notification service test

+++ src/main/java/com/salon/booking/service/NotificationService.java
@@ -11,5 +11,5 @@ public interface NotificationService {
 
     List<Notification> findAllRead(Integer userId);
 
-    long updateNotificationsReturningCount(Integer userId, LocalDateTime currentTime);
+    int updateNotificationsReturningCount(Integer userId, LocalDateTime currentTime);
 }
